- split autodetect.c into two .c files to make unit testing easier
