+++ MainAVR/src/autodetect.h
@@ -1,15 +1,9 @@
 #ifndef AUTODETECT_H
 #define AUTODETECT_H
 
-#include "settings.h"	// LDPType
-
 // detects whether the active player is LD-V1000 or PR-7820 and starts are detected player's main loop
 void ldv1000_or_pr7820_main_loop();
 void probably_ldv1000_mode();
 void probably_pr7820_mode();
-uint8_t IsPin11Raised();
-
-// detects which type mode to use when the user has selected 'other'
-LDPType detect_other_mode();
 
 #endif // AUTODETECT_H
